requirements.txt
matplotlib==3.5.1
config==0.5.1
hydra_core=1.1.1
numpy==1.21.2
omegaconf==2.1.1
opencv_python==4.5.5.64
Pillow==9.4.0
PyYAML==6.0
requests==2.27.1
scikit_learn==1.2.2
scipy==1.8.0
tensorboard==2.8.0
tensorboardX==2.6
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.0
git+https://github.com/wbaek/theconf.git
git+https://github.com/ildoonet/pytorch-gradual-warmup-lr.git
